% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_central_matches.R
\docType{data}
\name{central_matches}
\alias{central_matches}
\title{central_matches}
\format{data.frame}
\source{
\url{"https://raw.githubusercontent.com/JeffSackmann/tennis_atp/master/"}
}
\usage{
central_matches
}
\description{
The table of all matches
}
\author{
Leonard \email{leonard.fleutot@polytechnique.edu}
}
\keyword{data}
