% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{data_fill}
\alias{data_fill}
\title{Fills the blanks in the dataset}
\usage{
data_fill(central_matches)
}
\arguments{
\item{central_matches}{dataset of matches}
}
\description{
Fills the Nones in the surface collumn for the Davis Cup 2017
}
