% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{elo_calculation}
\alias{elo_calculation}
\title{Computes the elo}
\usage{
elo_calculation(central_matches)
}
\arguments{
\item{central_matches}{dataset of matches}
}
\description{
Basic function to compute the elo from the dataset of matches
}
