% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourneys.R
\name{filter_tourney}
\alias{filter_tourney}
\title{Filter Tennis Data on Tournament}
\usage{
filter_tourney(x, ...)
}
\arguments{
\item{x}{A data frame from the \code{tennis} package.}

\item{...}{Either the name or the id of the tournament(s). Might
be used together. The ids are formed by the last three characters
of \code{tourney_id}.}

\item{x}{A data frame from the \code{tennis} package.}

\item{...}{A vector with the ids and/or tournaments.}
}
\value{
A subset of \code{x} with the matches of the requested tournaments.
}
\description{
Select one or more tournaments from \code{x}.
}
\examples{
# find all the Grand Slam matches in 2016
library(dplyr)
atp_matches \%>\% filter_time(2016) \%>\%
    filter_tourney("Australian Open", "Roland Garros", "Wimbledon", "US Open")

find all the Grand Slam matches from 2010 to 2015
atp_matches \%>\% filter_time(2010, 2015) \%>\%
   filter_tourney(520, 540, 560, 580)

find all matches at Sydney, Brisbane and Australian Open
atp_matches \%>\% filter_tourney("Sydney", "Brisbane", 580)
}
