% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{find_player}
\alias{find_player}
\title{Find all the matches of a player}
\usage{
find_player(x, ..., won_lost = c("both", "won", "lost"))
}
\arguments{
\item{x}{A data frame from the \code{tennis} package.}

\item{...}{Character string of the full player name(s).
Multiple player names are separated by a comma.}

\item{won_lost}{Character string indicating if only the won matches,
lost matches, or both should be returned.}
}
\value{
A subset of \code{x} with the matches of the requested players.
}
\description{
Query function to return the matches of one or more players in the data set.
Function is case-insensitive, all names will be set to lower before matching.
}
\examples{
library(dplyr)
atp_matches \%>\% find_player("roger federer", "Rafael Nadal", "NOVAK DJOKOVIC")
}
