% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis}
\alias{tennis}
\title{Data for 14 tennis matches from 2010 to 2023}
\format{
\describe{
\item{\verb{Player 1}}{The first opponent in the match}
\item{\verb{Player 2}}{The second opponent in the match}
\item{\verb{Player 1 Wins}}{Says ‘yes’ if player 1 won the match and ‘no’ if player 2 won the match}
\item{\verb{Player 1 Ace}}{How many aces player 1 served in the match}
\item{\verb{Player 2 Ace}}{How many aces player 2 served in the match}
\item{\verb{Difference in Ace}}{What is the difference in the number of aces between each player}
\item{Year}{The year the match took place}
}
}
\source{
\url{https://www.ultimatetennisstatistics.com/tournament?tournamentId=19&tab=events}
}
\usage{
tennis
}
\description{
Data for 14 tennis matches from 2010 to 2023
}
\keyword{datasets}
