% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{tennis_reshape}
\alias{tennis_reshape}
\title{Reshape the tennis data from the perspective of one player}
\usage{
tennis_reshape(x, ...)
}
\arguments{
\item{x}{A data frame from the \code{tennis} package.}

\item{...}{Character string of the full player name(s).
Multiple player names are separated by a comma. If empty,
all players in the set will be retained (and the data set will
double in rows).}
}
\value{
For each match played by a requested player a record is created. The
columns \code{player}, \code{won_lost}, and \code{opponent} replace the
columns \code{winner_name} and \code{loser_name}. All the other columns
containing player information are renamed to either player_ or opponent_
(and p_ or o_) depending to which of the players they belong.
}
\description{
Each match will be converted  so its from the perspective of one player. When
both players of a match are queried, two records will be created from one
match. The reshaped data makes it easier to analyse the results of a single
player. Function is case-insensitive.
}
